refactor mock entitlement builder to separate file
